Replace static Iq::isXyzIq() with generic IQ type matching
Introducing QXmpp::isIqType<T>(QDomElement) which does the same. It
relys on a static PayloadXmlTag member in the IQ class.

As a fallback also `::checkIqType()` is used. This is necessary
especially for the QXmppDiscoveryIq which implements two different IQs
(disco info and disco items, which have different namespaces).

And also T::PayloadType::XmlTag works which should be very useful in the
future with a generic IQ<Payload> type, but of course can be used
already now.